refactor: Consolidate context menu command processing and update package structure
