+++ src/main/kotlin/dev/slne/discord/discord/interaction/command/context/commands/SendReplyDeadlineContextMenuCommand.kt
@@ -1,9 +1,9 @@
-package dev.slne.discord.discord.interaction.context.commands
+package dev.slne.discord.discord.interaction.command.context.commands
 
 import dev.slne.discord.annotation.DiscordContextMenuCommandMeta
 import dev.slne.discord.discord.interaction.command.commands.ticket.sendReplyDeadlineToChannel
-import dev.slne.discord.discord.interaction.context.DiscordContextMenuCommand
-import dev.slne.discord.discord.interaction.context.DiscordContextMenuCommandType
+import dev.slne.discord.discord.interaction.command.context.DiscordContextMenuCommand
+import dev.slne.discord.discord.interaction.command.context.DiscordContextMenuCommandType
 import dev.slne.discord.guild.permission.CommandPermission
 import net.dv8tion.jda.api.entities.User
 import net.dv8tion.jda.api.events.interaction.command.GenericContextInteractionEvent
