+++ src/main/resources/messages.properties
@@ -303,4 +303,5 @@ ticket.close.inactive.close-reason=Das Ticket wurde aufgrund von Inaktivität aut
 common.yes=Ja
 common.no=Nein
 common.waiting.for.ping=Warte auf Pings...
+interaction.context.menu.ticket-admin-panel.not-a-member={0} ist nicht in diesem Thread!
